Revert "Sync to da9da17db2c68140ad873d2701c3c05162ca8443"

+++ mobile/android/tests/background/junit4/src/org/mozilla/gecko/sync/repositories/android/VisitsHelperTest.java
@@ -85,9 +85,9 @@ public class VisitsHelperTest {
             for (int i = 0; i < recentVisits.size(); i++) {
                 JSONObject v = (JSONObject) recentVisits.get(i);
                 Long date = (Long) v.get("date");
-                Long type = (Long) v.get("type");
+                Integer type = (Integer) v.get("type");
                 Assert.assertEquals(Long.valueOf(baseDate - i * 100), date);
-                Assert.assertEquals(Long.valueOf(1), type);
+                Assert.assertEquals(Integer.valueOf(1), type);
             }
         } finally {
             provider.shutdown();
